+++ qoqo/src/measurements/cheated_basis_rotation_measurement.rs
@@ -65,8 +65,7 @@ impl CheatedPauliZProductWrapper {
             for c in circuits.into_iter() {
                 let tmp_c = CircuitWrapper::from_pyany(c.bind(py)).map_err(|err| {
                     PyTypeError::new_err(format!(
-                        "`circuits` argument is not a list of qoqo Circuits: {}",
-                        err
+                        "`circuits` argument is not a list of qoqo Circuits: {err}"
                     ))
                 })?;
                 new_circuits.push(tmp_c)
@@ -76,8 +75,7 @@ impl CheatedPauliZProductWrapper {
                 Some(c) => {
                     let tmp_c = CircuitWrapper::from_pyany(c.bind(py)).map_err(|err| {
                         PyTypeError::new_err(format!(
-                            "`constant_circuit` argument is not None or a qoqo Circuit: {}",
-                            err
+                            "`constant_circuit` argument is not None or a qoqo Circuit: {err}"
                         ))
                     })?;
                     Some(tmp_c)
@@ -86,8 +84,7 @@ impl CheatedPauliZProductWrapper {
             let input =
                 CheatedPauliZProductInputWrapper::from_pyany(input.bind(py)).map_err(|err| {
                     PyTypeError::new_err(format!(
-                        "`input` argument is not a qoqo CheatedInput: {}",
-                        err
+                        "`input` argument is not a qoqo CheatedInput: {err}"
                     ))
                 })?;
             Ok(Self {
@@ -144,8 +141,7 @@ impl CheatedPauliZProductWrapper {
             .evaluate(bit_registers, float_registers, complex_registers)
             .map_err(|x| {
                 PyRuntimeError::new_err(format!(
-                    "Error evaluating cheated PauliZ product measurement {:?}",
-                    x
+                    "Error evaluating cheated PauliZ product measurement {x:?}"
                 ))
             })
     }
@@ -204,10 +200,7 @@ impl CheatedPauliZProductWrapper {
                 .internal
                 .substitute_parameters(substituted_parameters)
                 .map_err(|x| {
-                    PyRuntimeError::new_err(format!(
-                        "Error substituting symbolic parameters {:?}",
-                        x
-                    ))
+                    PyRuntimeError::new_err(format!("Error substituting symbolic parameters {x:?}"))
                 })?,
         })
     }
@@ -391,8 +384,7 @@ impl CheatedPauliZProductWrapper {
             })?;
             deserialize(&bytes[..]).map_err(|err| {
                     PyTypeError::new_err(format!(
-                    "Python object cannot be converted to qoqo CheatedPauliZProduct: Deserialization failed: {}",
-                    err
+                    "Python object cannot be converted to qoqo CheatedPauliZProduct: Deserialization failed: {err}"
                 ))
                 })
         }