+++ roqoqo/src/quantum_program.rs
@@ -246,7 +246,7 @@ impl Display for QuantumProgram {
             }
         }
 
-        write!(f, "{}", s)
+        write!(f, "{s}")
     }
 }