+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This changelog track changes to the qoqo project starting at version v0.5.0
 
 ## Unreleased
 
+* Added the `number_of_qubits` method to the `Circuit` class.
 * Fixed bug omitting a factor of 1/2 when adding dephasing.
 * Added MultiQubitCNOT.